full pipeline working
