+++ tools/post_processing/min_e.py
@@ -12,7 +12,6 @@
 
 import matrix_element
 import minimize
-
 import argparse
 import csv
 import glob
