+++ tools/test_post_process.py
@@ -6,18 +6,18 @@ import os
 def main(args):
     os.chdir(args.path)
     # Obtain the differences in the sbias for each transition
-    diff_s_bias.get_diff_sbias()
+    #diff_s_bias.get_diff_sbias()
     # Obtain the average sbias for each bonding state
-    avg_s_bias.get_avg_sbias(diff_sbias_csv="diff_s_bias.csv", structure_sim_json=args.json)
+    #avg_s_bias.get_avg_sbias(diff_sbias_csv="diff_s_bias.csv", structure_sim_json=args.json)
     # Obtain the mfpt for each bonding state
-    mfpt.get_mfpt()
+    #mfpt.get_mfpt()
     # put together the mfpts in one file
-    mfpt.process_mfpts()
+    mfpt.compile_mfpts()
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--path", help="results_path", default='test_st_OGVarrc')
-    parser.add_argument("--json", help="json_path", default='../test_st_2.json')
+    parser.add_argument("--path", help="results_path", default='test_st')
+    #parser.add_argument("--json", help="json_path", default='../test_st.json')
     args = parser.parse_args()
     main(args)
